+++ stubs/app/Providers/SettingsServiceProvider.php
@@ -156,8 +156,8 @@ class SettingsServiceProvider extends ServiceProvider
             if (isset($fileManager['max_size_mb'])) {
                 config(['file-manager.settings.max_size_mb' => (int) $fileManager['max_size_mb']]);
             }
-            if (isset($fileManager['storage_quota_mb'])) {
-                config(['file-manager.settings.storage_quota_mb' => (int) $fileManager['storage_quota_mb']]);
+            if (isset($fileManager['storage_quota_gb'])) {
+                config(['file-manager.settings.storage_quota_gb' => (int) $fileManager['storage_quota_gb']]);
             }
             if (isset($fileManager['accepted_mimes'])) {
                 $mimes = $fileManager['accepted_mimes'];