+++ stubs/resources/js/pages/Admin/Settings/components/FileManagerTab.vue
@@ -9,7 +9,7 @@
     interface Props {
         settings: {
             max_size_mb: number;
-            storage_quota_mb: number;
+            storage_quota_gb: number;
             accepted_mimes: string[];
             allow_video: boolean;
             allow_audio: boolean;
@@ -26,7 +26,7 @@
             .cardSubtitle('sk-setting.file_manager.subtitle')
             .initialData({
                 max_size_mb: props.settings.max_size_mb,
-                storage_quota_mb: props.settings.storage_quota_mb,
+                storage_quota_gb: props.settings.storage_quota_gb,
                 accepted_mimes: props.settings.accepted_mimes,
                 allow_video: props.settings.allow_video,
                 allow_audio: props.settings.allow_audio,
@@ -38,7 +38,7 @@
             })
             .addFields(
                 FB.inputNumber().key('max_size_mb').min(1).max(1024).class('col-span-1'),
-                FB.inputNumber().key('storage_quota_mb').min(1).max(1048576).class('col-span-1'),
+                FB.inputNumber().key('storage_quota_gb').min(1).max(1024).class('col-span-1'),
                 FB.multiselect().key('accepted_mimes').class('col-span-full'),
                 FB.toggleSwitch().key('allow_video').label(false).class('col-span-1'),
                 FB.toggleSwitch().key('allow_audio').label(false).class('col-span-1'),